# mw-chat/docs/swagger.yaml
basePath: /chat
definitions:
  schemas.CreateMessagePayload:
    properties:
      message:
        type: string
      roomId:
        type: string
    required:
    - message
    - roomId
    type: object
  schemas.CreateMessageResponse:
    properties:
      message:
        $ref: '#/definitions/schemas.MessageResponse'
      users:
        items:
          type: string
        type: array
    required:
    - message
    - users
    type: object
  schemas.CreateRoomPayload:
    properties:
      name:
        type: string
        x-nullable: true
      roomType:
        type: string
      userId:
        type: string
        x-nullable: true
    required:
    - roomType
    type: object
  schemas.GetChatPreviewResponse:
    properties:
      unreadMessagesAmount:
        type: integer
    required:
    - unreadMessagesAmount
    type: object
  schemas.GetRoomsResponse:
    properties:
      rooms:
        items:
          $ref: '#/definitions/schemas.RoomPreviewResponse'
        type: array
      size:
        type: integer
    required:
    - rooms
    - size
    type: object
  schemas.MessageReader:
    properties:
      readDate:
        type: string
      userId:
        type: string
    required:
    - readDate
    - userId
    type: object
  schemas.MessageResponse:
    properties:
      message:
        type: string
      messageId:
        type: string
      messageReaders:
        items:
          $ref: '#/definitions/schemas.MessageReader'
        type: array
      ownerId:
        type: string
    required:
    - message
    - messageId
    - messageReaders
    - ownerId
    type: object
  schemas.RoomPopulatedResponse:
    properties:
      isBlocked:
        type: boolean
      messages:
        items:
          $ref: '#/definitions/schemas.MessageResponse'
        type: array
      name:
        type: string
        x-nullable: true
      roomId:
        type: string
      roomType:
        type: string
      users:
        items:
          $ref: '#/definitions/schemas.UserResponse'
        type: array
    required:
    - isBlocked
    - messages
    - name
    - roomId
    - roomType
    - users
    type: object
  schemas.RoomPreviewResponse:
    properties:
      isBlocked:
        type: boolean
      name:
        type: string
        x-nullable: true
      roomId:
        type: string
      roomType:
        type: string
      users:
        items:
          $ref: '#/definitions/schemas.UserResponse'
        type: array
    required:
    - isBlocked
    - name
    - roomId
    - roomType
    - users
    type: object
  schemas.UpdateMessageStatusPayload:
    properties:
      isRead:
        type: boolean
    required:
    - isRead
    type: object
  schemas.UserResponse:
    properties:
      role:
        type: string
      userId:
        type: string
    required:
    - role
    - userId
    type: object
info:
  contact: {}
  title: Masters way chat API
  version: "1.0"
paths:
  /dev/reset-db:
    get:
      description: resets db
      responses:
        "204":
          description: No Content
      summary: resets db
      tags:
      - dev
  /messages:
    post:
      consumes:
      - application/json
      operationId: create-message
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateMessagePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.CreateMessageResponse'
      summary: Create message
      tags:
      - message
  /messages/{messageId}/message-status:
    patch:
      consumes:
      - application/json
      description: Update message status by message Id
      operationId: update-message-status
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateMessageStatusPayload'
      - description: message Id
        in: path
        name: messageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Update message status
      tags:
      - message
  /rooms:
    post:
      consumes:
      - application/json
      operationId: create-room
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateRoomPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.RoomPopulatedResponse'
      summary: Create room for user
      tags:
      - room
  /rooms/{roomId}:
    get:
      consumes:
      - application/json
      operationId: get-room-by-id
      parameters:
      - description: room Id
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.RoomPopulatedResponse'
      summary: Get room by id
      tags:
      - room
    patch:
      consumes:
      - application/json
      operationId: update-room
      parameters:
      - description: room Id
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.RoomPopulatedResponse'
      summary: Update room for user
      tags:
      - room
  /rooms/{roomId}/users/{userId}:
    delete:
      consumes:
      - application/json
      operationId: delete-user-from-room
      parameters:
      - description: room Id
        in: path
        name: roomId
        required: true
        type: string
      - description: user Id to delete
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete user from room
      tags:
      - room
    post:
      consumes:
      - application/json
      operationId: add-user-to-room
      parameters:
      - description: room Id
        in: path
        name: roomId
        required: true
        type: string
      - description: user Id to delete
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.RoomPreviewResponse'
      summary: Add user to room
      tags:
      - room
  /rooms/list/{roomType}:
    get:
      consumes:
      - application/json
      operationId: get-rooms
      parameters:
      - description: 'room type: private | group'
        enum:
        - private
        - group
        in: path
        name: roomType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.GetRoomsResponse'
      summary: Get rooms for user
      tags:
      - room
  /rooms/preview:
    get:
      consumes:
      - application/json
      operationId: get-chat preview
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.GetChatPreviewResponse'
      summary: Get chat preview
      tags:
      - room
swagger: "2.0"